Move OpenAI code editor to intermediate layer
