+++ src/adapters/adapter_openai.cpp
@@ -1,8 +1,5 @@
 #include "adapter_openai.hpp"
 
-#include "structured_output.hpp"
-#include "system_prompts.hpp"
-
 #include <fmt/core.h>
 #include <stdexcept>
 
@@ -58,39 +55,6 @@ std::string OpenAI::query_responses_api_(const std::string &post_fields)
     }
 
     return response;
-}
-
-std::expected<OpenAIEdit, OpenAIError> OpenAI::query_edit_code(const std::string &prompt)
-{
-    const nlohmann::json response_format = {
-        {
-            "format",
-            {
-                { "name", "updated_code" },
-                { "schema", structured_output::schema_edit_code() },
-                { "strict", true },
-                { "type", "json_schema" },
-            },
-        }
-    };
-    const nlohmann::json fields = {
-        { "input", prompt },
-        { "instructions", system_prompts::system_prompt_edit_code() },
-        { "model", this->model_ },
-        { "store", false },
-        { "temperature", 1.00 },
-        { "text", response_format },
-    };
-
-    const std::string response = this->query_responses_api_(fields.dump());
-
-    long http_status_code = this->get_http_status_code_();
-    if (http_status_code != 200) {
-        return std::unexpected(OpenAIError(response, http_status_code));
-    }
-
-    const double total_time = this->get_rtt_time_();
-    return OpenAIEdit(response, total_time);
 }
 
 } // namespace adapters
